# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CEREBRO-RED v2 API",
    "description": "\n    # Autonomous Local LLM Red Teaming Suite - Research Edition\n    \n    ## Overview\n    CEREBRO-RED v2 implements the PAIR (Prompt Automatic Iterative Refinement) algorithm\n    for automated vulnerability discovery in Large Language Models.\n    \n    ## Features\n    - **Multi-Provider LLM Support**: Ollama, Azure OpenAI, OpenAI\n    - **8 Attack Strategies**: Obfuscation, Context Flooding, Roleplay Injection, etc.\n    - **LLM-as-a-Judge**: Semantic evaluation with 7 criteria\n    - **Real-Time Progress**: WebSocket streaming for live updates\n    - **Research-Grade Telemetry**: JSONL audit logs for analysis\n    \n    ## Authentication\n    All endpoints (except `/health`, `/docs`, `/metrics`) require API key authentication via `X-API-Key` header.\n    \n    ## Rate Limiting\n    Default: 60 requests per minute per IP address.\n    \n    ## References\n    - PAIR Paper: https://arxiv.org/abs/2310.08419\n    - GitHub: https://github.com/your-org/cerebro-red-v2\n    ",
    "version": "2.0.0",
    "contact": {
      "name": "CEREBRO-RED v2 Support",
      "url": "https://github.com/your-org/cerebro-red-v2/issues",
      "email": "support@cerebro-red.example.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "paths": {
...
    "/api/v1/scan/start": {
      "post": {
        "tags": [
          "scans"
        ],
        "summary": "Start Scan",
        "description": "Start a new scan (experiment execution).\n\nRuns experiment in background and returns immediately with experiment_id.\nUse WebSocket endpoint or status endpoint to track progress.",
        "operationId": "start_scan_api_v1_scan_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/scan/{experiment_id}/pause": {
      "post": {
        "tags": [
          "scans"
        ],
        "summary": "Pause Scan",
        "description": "Pause a running scan.",
        "operationId": "pause_scan_api_v1_scan__experiment_id__pause_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Experiment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scan/{experiment_id}/resume": {
      "post": {
        "tags": [
          "scans"
        ],
        "summary": "Resume Scan",
        "description": "Resume a paused scan.",
        "operationId": "resume_scan_api_v1_scan__experiment_id__resume_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Experiment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scan/{experiment_id}/cancel": {
      "post": {
        "tags": [
          "scans"
        ],
        "summary": "Cancel Scan",
        "description": "Cancel a running scan.\n\nImmediately cancels the running task and updates status to FAILED.",
        "operationId": "cancel_scan_api_v1_scan__experiment_id__cancel_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Experiment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scan/status/{experiment_id}": {
      "get": {
        "tags": [
          "scans"
        ],
        "summary": "Get Scan Status",
        "description": "Get current scan status and progress.",
        "operationId": "get_scan_status_api_v1_scan_status__experiment_id__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Experiment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scan/batch": {
      "post": {
        "tags": [
          "scans"
        ],
        "summary": "Start Batch Scan",
        "description": "Start batch scan (multiple experiments in parallel).\n\nEach experiment in the request is executed independently with its own config and prompts.\nExperiments run concurrently using asyncio.create_task within a single background task.",
        "operationId": "start_batch_scan_api_v1_scan_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchScanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchScanRequest": {
        "properties": {
          "experiments": {
            "items": {
              "$ref": "#/components/schemas/ExperimentConfig"
            },
            "type": "array",
            "title": "Experiments"
          }
        },
        "type": "object",
        "required": [
          "experiments"
        ],
        "title": "BatchScanRequest",
        "description": "Request to start batch scans."
      },
      "ExperimentConfig": {
        "properties": {
          "experiment_id": {
            "type": "string",
            "format": "uuid",
            "title": "Experiment Id",
            "description": "Unique experiment ID"
          },
          "name": {
            "type": "string",
            "maxLength": 200,
            "title": "Name",
            "description": "Experiment name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Experiment description"
          },
          "target_model_provider": {
            "$ref": "#/components/schemas/LLMProvider",
            "description": "Target model provider"
          },
          "target_model_name": {
            "type": "string",
            "title": "Target Model Name",
            "description": "Target model name"
          },
          "attacker_model_provider": {
            "$ref": "#/components/schemas/LLMProvider",
            "description": "Attacker model provider"
          },
          "attacker_model_name": {
            "type": "string",
            "title": "Attacker Model Name",
            "description": "Attacker model name"
          },
          "judge_model_provider": {
            "$ref": "#/components/schemas/LLMProvider",
            "description": "Judge model provider"
          },
          "judge_model_name": {
            "type": "string",
            "title": "Judge Model Name",
            "description": "Judge model name"
          },
          "initial_prompts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Initial Prompts",
            "description": "Initial prompts to test"
          },
          "strategies": {
            "items": {
              "$ref": "#/components/schemas/AttackStrategyType"
            },
            "type": "array",
            "minItems": 1,
            "title": "Strategies",
            "description": "Attack strategies to use"
          },
          "max_iterations": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Max Iterations",
            "description": "Maximum iterations",
            "default": 20
          },
          "max_concurrent_attacks": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Max Concurrent Attacks",
            "description": "Maximum concurrent attacks",
            "default": 5
          },
          "success_threshold": {
            "type": "number",
            "maximum": 10.0,
            "minimum": 0.0,
            "title": "Success Threshold",
            "description": "Success threshold score",
            "default": 7.0
          },
          "timeout_seconds": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Timeout Seconds",
            "description": "Experiment timeout in seconds",
            "default": 3600
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Creation timestamp"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata",
            "description": "Additional metadata"
          }
        },
        "type": "object",
        "required": [
          "name",
          "target_model_provider",
          "target_model_name",
          "attacker_model_provider",
          "attacker_model_name",
          "judge_model_provider",
          "judge_model_name",
          "initial_prompts",
          "strategies"
        ],
        "title": "ExperimentConfig",
        "description": "Experiment configuration model."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LLMProvider": {
        "type": "string",
        "enum": [
          "ollama",
          "azure",
          "openai"
        ],
        "title": "LLMProvider",
        "description": "LLM provider types."
      },
      "ScanStartRequest": {
        "properties": {
          "experiment_config": {
            "$ref": "#/components/schemas/ExperimentConfig"
          }
        },
        "type": "object",
        "required": [
          "experiment_config"
        ],
        "title": "ScanStartRequest",
        "description": "Request to start a new scan."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:9000/api/v1",
...
    },
    {
...
    }
  ]
}